"feat: adjusts for match and attempts"

+++ backend/src/main/java/com/case_fullstack/mastermind/repositories/AttemptRepository.java
@@ -3,5 +3,5 @@ package com.case_fullstack.mastermind.repositories;
 import com.case_fullstack.mastermind.models.entities.Attempt;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AttemptRepository extends JpaRepository<Attempt, Integer> {
+public interface AttemptRepository extends JpaRepository<Attempt, Long> {
 }
